requirements.txt
Django==4.0.5
djangorestframework==3.12.4
django-filter==2.4.0
Markdown==3.1.1
psycopg2-binary==2.8.4
drf-yasg==1.21.0
gunicorn==19.9.0
celery==5.1.0
ipython==7.34.0
cryptography==2.8.0
ffmpy==0.3.0
django-cors-headers==3.7.0
ffmpeg-python==0.2.0
xlrd==1.2.0
openpyxl==3.0.3
django-json-widget==1.1.1
attrs==21.4.0
sentry-sdk==1.0.0
django-fieldsets-with-inlines==0.6
boto3==1.24.13
factory-boy==3.2.0
django-constance==2.8.0
django-picklefield==3.0.1
django-celery-results==2.2.0
Pillow==7.1.1
coverage==6.2
pytest-cov==3.0.0
unittest-xml-reporting==3.2.0
pytest-django==4.5.2
XlsxWriter==3.0.3
mypy==0.961
python-dotenv==0.20.0
matplotlib==3.5.1
numpy==1.22.3
pathlib==1.0.1
scipy==1.8.0
requests==2.27.1
types-requests==2.28.0
django-timestampedmodel==0.1.0